Implementação Inicial REST

+++ src/main/java/br/com/tokiomarine/seguradora/avaliacao/repository/EstudanteRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import br.com.tokiomarine.seguradora.avaliacao.entidade.Estudante;
 
 @Repository
-public interface EstudanteRepository extends CrudRepository<Estudante, Integer>  {
+public interface EstudanteRepository extends CrudRepository<Estudante, Long>  {
 
 	List<Estudante> findByNome(String nome);
